refactor(registry-manager): Make registry manager mockable
