+++ iothub/service/tests/RegistryManagerTests.cs
@@ -45,7 +45,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.GetAsync<Device>(It.IsAny<Uri>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), null, false, It.IsAny<CancellationToken>())).ReturnsAsync(deviceToReturn);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             var device = await registryManager.GetDeviceAsync(DeviceId).ConfigureAwait(false);
             Assert.AreSame(deviceToReturn, device);
             restOpMock.VerifyAll();
@@ -56,7 +56,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task GetDeviceAsyncWithNullDeviceIdTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.GetDeviceAsync(null).ConfigureAwait(false);
             Assert.Fail("Calling GetDeviceAsync with null device id did not throw an exception.");
         }
@@ -75,7 +75,7 @@ namespace Microsoft.Azure.Devices.Api.Test
                 true,
                 It.IsAny<CancellationToken>())).ReturnsAsync(devicesToReturn);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 
 #pragma warning disable CS0618 // Type or member is obsolete
             var returnedDevices = await registryManager.GetDevicesAsync(1).ConfigureAwait(false);
@@ -93,7 +93,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PutAsync(It.IsAny<Uri>(), It.IsAny<Device>(), It.IsAny<PutOperationType>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<CancellationToken>())).ReturnsAsync(deviceToReturn);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             var returnedDevice = await registryManager.AddDeviceAsync(deviceToReturn).ConfigureAwait(false);
             Assert.AreSame(deviceToReturn, returnedDevice);
             restOpMock.VerifyAll();
@@ -105,7 +105,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         {
             var deviceToReturn = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected, ETag = "123" };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDeviceAsync(deviceToReturn).ConfigureAwait(false);
             Assert.Fail("RegisterDevice API did not throw exception when bad deviceid was used.");
         }
@@ -116,7 +116,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         {
             var deviceToReturn = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "123" };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDeviceAsync(deviceToReturn).ConfigureAwait(false);
             Assert.Fail("RegisterDevice API did not throw exception when ETag was set.");
         }
@@ -126,7 +126,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task RegisterDeviceAsyncWithNullDeviceTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDeviceAsync(null).ConfigureAwait(false);
             Assert.Fail("RegisterDevice API did not throw exception when the device parameter was null.");
         }
@@ -136,7 +136,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task RegisterDeviceAsyncWithDeviceIdNullTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDeviceAsync(new Device()).ConfigureAwait(false);
             Assert.Fail("RegisterDevice API did not throw exception when the device's id was not set.");
         }
@@ -149,7 +149,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             // '/' is not a valid character in DeviceId
             var badDevice = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.AddDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -164,7 +164,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             // '/' is not a valid character in DeviceId
             var badDevice = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("RegisterDevices API did not throw exception when bad deviceid was used.");
         }
@@ -176,7 +176,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device("234") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.AddDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -190,7 +190,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device("234") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("RegisterDevices API did not throw exception when ETag was used.");
         }
@@ -202,7 +202,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             Device badDevice = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.AddDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -216,7 +216,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             Device badDevice = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("RegisterDevices API did not throw exception when Null device was used.");
         }
@@ -226,7 +226,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task RegisterDevicesAsyncWithNullDeviceListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.AddDevicesAsync(new List<Device>()).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -238,7 +238,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task RegisterDevices2AsyncWithNullDeviceListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDevices2Async(new List<Device>()).ConfigureAwait(false);
             Assert.Fail("RegisterDevices API did not throw exception when Null device list was used.");
         }
@@ -250,7 +250,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.AddDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -264,7 +264,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.AddDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("RegisterDevices API did not throw exception when deviceId was null.");
         }
@@ -276,7 +276,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PutAsync(It.IsAny<Uri>(), It.IsAny<Device>(), It.IsAny<PutOperationType>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<CancellationToken>())).ReturnsAsync(deviceToReturn);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             var returnedDevice = await registryManager.UpdateDeviceAsync(deviceToReturn).ConfigureAwait(false);
             Assert.AreSame(deviceToReturn, returnedDevice);
             restOpMock.VerifyAll();
@@ -293,7 +293,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task UpdateDeviceWithNullDeviceTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDeviceAsync(null).ConfigureAwait(false);
             Assert.Fail("UpdateDevice api did not throw exception when the device parameter was null.");
         }
@@ -303,7 +303,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task UpdateDeviceWithDeviceIdNullTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDeviceAsync(new Device() { ETag = "*" }).ConfigureAwait(false);
             Assert.Fail("UpdateDevice api did not throw exception when the device's id was null.");
         }
@@ -313,7 +313,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task UpdateDeviceWithInvalidDeviceIdTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             // '/' is not a valid char in DeviceId
             await registryManager.UpdateDeviceAsync(new Device("/baddevice") { ETag = "*" }).ConfigureAwait(false);
             Assert.Fail("UpdateDevice api did not throw exception when the deviceid was invalid.");
@@ -326,7 +326,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -340,7 +340,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("UpdateDevices API did not throw exception when bad deviceid was used.");
         }
@@ -352,7 +352,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device("234") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -366,7 +366,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device("234") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("UpdateDevices API did not throw exception when ETag was null.");
         }
@@ -378,7 +378,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             Device badDevice = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -392,7 +392,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             Device badDevice = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("UpdateDevices API did not throw exception when Null device was used.");
         }
@@ -402,7 +402,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task UpdateDevicesAsyncWithNullDeviceListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>()).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -414,7 +414,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task UpdateDevices2AsyncWithNullDeviceListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>()).ConfigureAwait(false);
             Assert.Fail("UpdateDevices API did not throw exception when Null device list was used.");
         }
@@ -426,7 +426,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -440,7 +440,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("UpdateDevices API did not throw exception when deviceId was null.");
         }
@@ -452,7 +452,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice2 = new Device("234") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<string[]>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<string[]>)null);
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { goodDevice1, goodDevice2 }, true, CancellationToken.None).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -466,7 +466,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<BulkRegistryOperationResult>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { goodDevice1, goodDevice2 }, true, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -479,7 +479,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<string[]>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<string[]>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { badDevice1, badDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -494,7 +494,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<BulkRegistryOperationResult>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { badDevice1, badDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -506,7 +506,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<string[]>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<string[]>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.UpdateDevicesAsync(new List<Device>() { goodDevice1, goodDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -520,7 +520,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<BulkRegistryOperationResult>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateDevices2Async(new List<Device>() { goodDevice1, goodDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -531,7 +531,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             var mockETag = new ETagHolder() { ETag = "*" };
             restOpMock.Setup(restOp => restOp.DeleteAsync(It.IsAny<Uri>(), mockETag, It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), null, It.IsAny<CancellationToken>())).Returns(Task.FromResult(0));
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDeviceAsync(new Device()).ConfigureAwait(false);
             restOpMock.VerifyAll();
         }
@@ -541,7 +541,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task DeleteDeviceAsyncWithNullIdTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDeviceAsync(string.Empty).ConfigureAwait(false);
             Assert.Fail("Delete API did not throw exception when the device id was null.");
         }
@@ -553,7 +553,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -567,7 +567,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected };
             var badDevice = new Device("/baddevice") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("DeleteDevices API did not throw exception when bad deviceid was used.");
         }
@@ -579,7 +579,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device("234") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -593,7 +593,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device("234") { ConnectionState = DeviceConnectionState.Connected };
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("DeleteDevices API did not throw exception when ETag was null.");
         }
@@ -605,7 +605,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             Device badDevice = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -619,7 +619,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             Device badDevice = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("DeleteDevices API did not throw exception when Null device was used.");
         }
@@ -629,7 +629,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task DeleteDevicesAsyncWithNullDeviceListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>()).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -641,7 +641,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task DeleteDevices2AsyncWithNullDeviceListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>()).ConfigureAwait(false);
             Assert.Fail("DeleteDevices API did not throw exception when Null device list was used.");
         }
@@ -653,7 +653,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -667,7 +667,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodDevice = new Device("123") { ConnectionState = DeviceConnectionState.Connected, ETag = "234" };
             var badDevice = new Device();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { goodDevice, badDevice }).ConfigureAwait(false);
             Assert.Fail("DeleteDevices API did not throw exception when deviceId was null.");
         }
@@ -680,7 +680,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<string[]>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<string[]>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { goodDevice1, goodDevice2 }, true, CancellationToken.None).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -694,7 +694,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<BulkRegistryOperationResult>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { goodDevice1, goodDevice2 }, true, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -707,7 +707,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<string[]>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<string[]>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { badDevice1, badDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -722,7 +722,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<BulkRegistryOperationResult>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { badDevice1, badDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -734,7 +734,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<string[]>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<string[]>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
 #pragma warning disable CS0618 // Type or member is obsolete
             await registryManager.RemoveDevicesAsync(new List<Device>() { goodDevice1, goodDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
 #pragma warning restore CS0618 // Type or member is obsolete
@@ -748,7 +748,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.PostAsync<IEnumerable<ExportImportDevice>, Task<BulkRegistryOperationResult>>(It.IsAny<Uri>(), It.IsAny<IEnumerable<ExportImportDevice>>(), It.IsAny<IDictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>>(), It.IsAny<IDictionary<string, string>>(), It.IsAny<CancellationToken>())).ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.RemoveDevices2Async(new List<Device>() { goodDevice1, goodDevice2 }, false, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -759,7 +759,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodTwin = new Twin("123");
             var badTwin = new Twin("/badTwin");
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { goodTwin, badTwin }).ConfigureAwait(false);
             Assert.Fail("UpdateTwins API did not throw exception when bad deviceid was used.");
         }
@@ -771,7 +771,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodTwin = new Twin("123") { ETag = "234" };
             var badTwin = new Twin("234");
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { goodTwin, badTwin }).ConfigureAwait(false);
             Assert.Fail("UpdateTwins API did not throw exception when ETag was null.");
         }
@@ -783,7 +783,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodTwin = new Twin("123") { ETag = "234" };
             Twin badTwin = null;
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { goodTwin, badTwin }).ConfigureAwait(false);
             Assert.Fail("UpdateTwins API did not throw exception when Null twin was used.");
         }
@@ -793,7 +793,7 @@ namespace Microsoft.Azure.Devices.Api.Test
         public async Task UpdateTwins2AsyncWithNullTwinListTest()
         {
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>()).ConfigureAwait(false);
             Assert.Fail("UpdateTwins API did not throw exception when Null twin list was used.");
         }
@@ -805,7 +805,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var goodTwin = new Twin("123") { ETag = "234" };
             var badTwin = new Twin();
             var restOpMock = new Mock<IHttpClientHelper>();
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { goodTwin, badTwin }).ConfigureAwait(false);
             Assert.Fail("UpdateTwins API did not throw exception when deviceId was null.");
         }
@@ -825,7 +825,7 @@ namespace Microsoft.Azure.Devices.Api.Test
                     It.IsAny<CancellationToken>()))
                 .ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { goodTwin1, goodTwin2 }, true, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -845,7 +845,7 @@ namespace Microsoft.Azure.Devices.Api.Test
                 It.IsAny<CancellationToken>()))
                 .ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { badTwin1, badTwin2 }, false, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -864,7 +864,7 @@ namespace Microsoft.Azure.Devices.Api.Test
                 It.IsAny<CancellationToken>()))
                 .ReturnsAsync((Task<BulkRegistryOperationResult>)null);
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.UpdateTwins2Async(new List<Twin>() { goodTwin1, goodTwin2 }, false, CancellationToken.None).ConfigureAwait(false);
         }
 
@@ -874,7 +874,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.Dispose());
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             registryManager.Dispose();
             restOpMock.Verify(restOp => restOp.Dispose(), Times.Once());
         }
@@ -885,7 +885,7 @@ namespace Microsoft.Azure.Devices.Api.Test
             var restOpMock = new Mock<IHttpClientHelper>();
             restOpMock.Setup(restOp => restOp.Dispose());
 
-            var registryManager = new HttpRegistryManager(restOpMock.Object, IotHubName);
+            var registryManager = new RegistryManager(IotHubName, restOpMock.Object);
             await registryManager.CloseAsync().ConfigureAwait(false);
             restOpMock.Verify(restOp => restOp.Dispose(), Times.Never());
         }
